+++ modules/statics/src/coins/solTokens.ts
@@ -3067,6 +3067,38 @@ export const solTokens = [
     UnderlyingAsset['sol:rekt'],
     SOL_TOKEN_FEATURES
   ),
+  solToken(
+    'cb3a07cd-0b6b-4e7c-810c-d93b257a7e34',
+    'sol:pipe',
+    'Pipe',
+    9,
+    '7s9MoSt7VV1J3jVNnw2AyocsQDBdCkPYz5apQDPKy9i5',
+    '7s9MoSt7VV1J3jVNnw2AyocsQDBdCkPYz5apQDPKy9i5',
+    UnderlyingAsset['sol:pipe'],
+    SOL_TOKEN_FEATURES
+  ),
+  solToken(
+    '9dbb51ba-de77-4317-a974-157f552646b5',
+    'sol:glxy',
+    'Galaxy Digital Inc. Class A Common Stock',
+    6,
+    '2HehXG149TXuVptQhbiWAWDjbbuCsXSAtLTB5wc2aajK',
+    '2HehXG149TXuVptQhbiWAWDjbbuCsXSAtLTB5wc2aajK',
+    UnderlyingAsset['sol:glxy'],
+    SOL_TOKEN_FEATURES,
+    ProgramID.Token2022ProgramId
+  ),
+  solToken(
+    'b44f7798-4c6f-4135-ab1f-501535641b9e',
+    'sol:ford',
+    'Forward Industries, Inc',
+    6,
+    '7GzQgf6DPo6ZANjnbhe9tNCpkGTv3zqHbsDx74jyQf9',
+    '7GzQgf6DPo6ZANjnbhe9tNCpkGTv3zqHbsDx74jyQf9',
+    UnderlyingAsset['sol:ford'],
+    SOL_TOKEN_FEATURES,
+    ProgramID.Token2022ProgramId
+  ),
   tsolToken(
     'b98c5a7a-49c5-45f1-a6ee-b08dff596a7d',
     'tsol:srm',