+++ modules/sdk-coin-etc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-etc@2.5.1...@bitgo/sdk-coin-etc@2.5.2) (2025-10-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-etc
+
 ## [2.5.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-etc@2.5.0...@bitgo/sdk-coin-etc@2.5.1) (2025-09-29)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-etc