+++ modules/sdk-coin-tao/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-tao@1.12.1...@bitgo/sdk-coin-tao@1.12.2) (2025-10-02)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-tao
+
 ## [1.12.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-tao@1.12.0...@bitgo/sdk-coin-tao@1.12.1) (2025-09-29)
 
 ### Bug Fixes